backend/requirements.txt
# =============================================================================
# YouTube Influencer Matching Agent - Python Dependencies
# =============================================================================
# このファイルには、バックエンドAPIで使用するPythonパッケージを定義

# -----------------------------------------------------------------------------
# Core Framework
# -----------------------------------------------------------------------------
# FastAPI - 高速なWeb APIフレームワーク
fastapi==0.104.1
# ASGI サーバー
uvicorn[standard]==0.24.0
# 本番用ASGIサーバー
gunicorn==21.2.0

# -----------------------------------------------------------------------------
# HTTP & API関連
# -----------------------------------------------------------------------------
# HTTPクライアント
httpx==0.25.2
requests==2.31.0
# API ドキュメント
pydantic==2.5.0
pydantic-settings==2.1.0

# -----------------------------------------------------------------------------
# Google Cloud SDKs
# -----------------------------------------------------------------------------
# Google Cloud Core
google-cloud-core==2.3.3
# Firestore
google-cloud-firestore==2.13.1
# BigQuery
google-cloud-bigquery==3.13.0
# Cloud Storage
google-cloud-storage==2.10.0
# Vertex AI
google-cloud-aiplatform==1.71.1
# Cloud Functions
functions-framework==3.5.0
# Cloud Monitoring
google-cloud-monitoring==2.16.0
# Cloud Logging
google-cloud-logging==3.8.0

# AI/ML関連
google-generativeai==0.3.2
vertexai==1.71.1

# -----------------------------------------------------------------------------
# 認証・セキュリティ
# -----------------------------------------------------------------------------
# Google Auth
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
# JWT処理
PyJWT[crypto]==2.8.0
# パスワードハッシュ化
passlib[bcrypt]==1.7.4
# OAuth
authlib==1.2.1

# -----------------------------------------------------------------------------
# データベース・ORM
# -----------------------------------------------------------------------------
# SQLAlchemy（将来的なRDB対応）
SQLAlchemy==2.0.23
# Alembic（マイグレーション）
alembic==1.12.1

# -----------------------------------------------------------------------------
# YouTube API
# -----------------------------------------------------------------------------
# Google API Client
google-api-python-client==2.108.0
google-api-core==2.12.0
# YouTube関連
youtube-transcript-api==0.6.1

# -----------------------------------------------------------------------------
# メール送信
# -----------------------------------------------------------------------------
# SendGrid
sendgrid==6.10.0
# SMTP（バックアップ）
aiosmtplib==3.0.1
# メールテンプレート
jinja2==3.1.2

# -----------------------------------------------------------------------------
# データ処理・分析
# -----------------------------------------------------------------------------
# pandas
pandas==2.1.3
# NumPy
numpy==1.25.2
# 日付処理
python-dateutil==2.8.2
# JSON処理
orjson==3.9.10

# -----------------------------------------------------------------------------
# 非同期処理・キュー
# -----------------------------------------------------------------------------
# Celery（タスクキュー）
celery[redis]==5.3.4
# Redis
redis==5.0.1
# 非同期処理（asyncioは標準ライブラリなので削除）
aiofiles==23.2.1

# -----------------------------------------------------------------------------
# 設定・環境変数
# -----------------------------------------------------------------------------
# .env ファイル処理
python-dotenv==1.0.0
# 設定管理
dynaconf==3.2.4

# -----------------------------------------------------------------------------
# ログ・監視
# -----------------------------------------------------------------------------
# 構造化ログ
structlog==23.2.0
# Rich（美しいログ出力）
rich==13.7.0
# プロメテウス メトリクス
prometheus-client==0.19.0

# -----------------------------------------------------------------------------
# テスト関連
# -----------------------------------------------------------------------------
# pytest
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
# HTTPテスト（上記で定義済みのためコメントアウト）
# httpx==0.25.2
# ファクトリボーイ
factory-boy==3.3.0

# -----------------------------------------------------------------------------
# 開発・デバッグ
# -----------------------------------------------------------------------------
# コード品質
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
# pre-commit
pre-commit==3.6.0

# -----------------------------------------------------------------------------
# ユーティリティ
# -----------------------------------------------------------------------------
# UUID生成
shortuuid==1.0.11
# 画像処理
Pillow==10.1.0
# URL処理
yarl==1.9.3
# 時間処理
arrow==1.3.0
# バリデーション
validators==0.22.0
# 文字列処理
fuzzywuzzy[speedup]==0.18.0

# -----------------------------------------------------------------------------
# 本番環境用
# -----------------------------------------------------------------------------
# APM（Application Performance Monitoring）
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-gcp-monitoring==1.6.0

# セキュリティスキャン
safety==2.3.5
bandit==1.7.5

# -----------------------------------------------------------------------------
# バージョン固定（セキュリティ対応）
# -----------------------------------------------------------------------------
# 脆弱性対応のため、特定バージョンに固定
cryptography>=41.0.0  # セキュリティアップデート
urllib3>=2.0.0        # セキュリティアップデート
certifi>=2023.7.22    # 証明書アップデート